added check to deleteItem resolver to ensure we are not deleting other peoples items

+++ src/resolvers/Mutation.js
@@ -43,7 +43,21 @@ const Mutations = {
   },
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
-    const item = await ctx.db.query.item({ where }, `{ id title }`);
+    const item = await ctx.db.query.item({ where }, `{ id title user { id } }`);
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx
+                            .request
+                            .user
+                            .permissions
+                            .some(
+                              permission => ['ADMIN', 'ITEMDELETE']
+                              .includes(permission)
+                            )
+    
+    if (!ownsItem || !hasPermission) {
+      throw new Error('You don\'t have permission to do that')
+    }
+
     return ctx.db.mutation.deleteItem({ where }, info);
   },
